refactor: 점수 계산 로직 수정
