+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,9 +1,6 @@
 package blackjack.domain
 
-data class Card(val suit: Suit, val cardNumber: CardNumber) {
-    val points: List<Int>
-        get() = cardNumber.points
-}
+data class Card(val suit: Suit, val cardNumber: CardNumber)
 
 enum class Suit {
     SPADE, HEART, DIAMOND, CLOVER
